feat(structure du code): sépare le code en dossiers (core, utils, data, sound...)
